% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_rasl.R
\name{multi_rasl}
\alias{multi_rasl}
\title{Do the extraction from raster images using las point data}
\usage{
multi_rasl(las_list, dsf_list, kwsindice, hmin)
}
\arguments{
\item{las_list}{a list of point cloud data}

\item{dsf_list}{a list of raster image data, mainly are tif image that read as raster}

\item{kwsindice}{the parameter for find tree}
}
\value{
a data list
}
\description{
multi_rasl is a function that used in the shiny app for data extraction,
it first use the point cloud data to do tree detection and tree crown polygon, and then
use this polygon to apply on the multispectal image to extract the multispecral
of each tree and then save as a data list
}
\examples{
none



}
