% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_alin.R
\name{raster_alin}
\alias{raster_alin}
\title{raster_alin aline single raster spectral images together as list}
\usage{
raster_alin(url)
}
\arguments{
\item{url}{the link that contain all raster images}
}
\value{
a list of raster images, used for multi_rasl function
}
\description{
raster_alin aline single raster spectral images together as list
}
\examples{
#please download the example data from this link:https://ln5.sync.com/dl/d6899c6f0/3g32725x-b85yuvm3-ba68kfre-jewun6fk
urll1 <- 'D:/Desktop/examples/'
images <- raster_alin(urll1)\%>\% unlist(recursive = F)\%>\% do.call(c,.)
images




}
