% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_old2.R
\name{SpecexR_app}
\alias{SpecexR_app}
\title{SpecexR_app
this is an shiny app mainly for multispectral extraction using LAS point cloud data}
\usage{
SpecexR_app()
}
\arguments{
\item{...}{}
}
\value{
final data with x,y location information
}
\description{
SpecexR_app
this is an shiny app mainly for multispectral extraction using LAS point cloud data
}
\examples{
 none

}
