% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_alin.R
\name{raster_alin}
\alias{raster_alin}
\title{raster_alin aline single raster spectral images together as list}
\usage{
raster_alin(url)
}
\arguments{
\item{url}{the link that contain all raster images}
}
\value{
a list of raster images, used for multi_rasl function
}
\description{
raster_alin aline single raster spectral images together as list
}
\examples{
urll1 <- 'E:/PCGSPRO_1623035349/17366648332/42/map/'
ras_all <- raster_alin(urll1)
ras_all <-ras_all \%>\% unlist(recursive = F)\%>\%  unlist(recursive = F)




}
